Refactor ProfileApplier Class
Rewrote ApplyProfile method to call two separate methods: SetNicAddress
and SetDNSServers to reflect the actual work being done.  Modified the
class to include two separate methods that are used to create netsh
arguments for DHCP and static configurations so that these argument
constructions could be tested with unit tests.

The Netsh argument creation methods are currently public to facilitate
testing, but I would prefere if they were private.  I could not figure out
a way to keep them private but expose them (in a resonable fashion) to the
testing project.  This may need to be a future refactoring.

Resolves: #5