Use Profile Command now Accepts NIC Name
The Use Profile command now takes two argumenst: a profile to apply and
the name of a NIC to apply it to.  If the profile name is not found then
it will apply a default DHCP configuration.  If the NIC is not found then
it will use the default NIC name of "Local Area Connection".

The "Local Area Connection" default will not work on Windows 10 machines
so the NIC name will have to be explicitly supplied for now.

Resolves: #4