Fixed issues with malformed, missing, or empty profile.xml file.
If the profiles.xml file was missing or malformed in any way the program
would throw an uncaught exception.  Now, an empty or missing file will
just return an empty collection of profiles.  A malformed file will still
throw an exception, but the program will catch that exception and prompt
the user to fix the file.

Resolves: #7