Check for and Create AppData Folder on Start
The profile file is automatically created in the SetIP folder of the %AppData% directory.  However, if
the SetIP directory did not exist an exception was thrown when trying to create the file. This change
adds code to check that the directory exists and then creates the directory if it is not found.

Resolves: #12