Add testing and start using CLImber
This is an intermittent commit. It shifts the unit test solution to
XUnit instead of MSTest. It also begins the process of updating the
unit tests to work with updated command classes.

Most commands have been updated to use CLImber, but this process has
stopped until unit tests can be written to verify functionality.